Backend/yolo-detection-ms/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
ultralytics==8.0.200
opencv-python==4.8.1.78
numpy==1.24.3
pydantic==2.4.2
python-dotenv==1.0.0

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.2
